LPS-72474 Avoid deleting latests approved version
In some corner cases, it is possible for the version not to change at
all between changes (e.g. originally the file entry had no contents,
and it is modified to contain a document but the other fields are not
updated).

In these cases, the file version will be updated in place, so no need
to remove or override anything.

+++ modules/apps/collaboration/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/exportimport/data/handler/FileEntryStagedModelDataHandler.java
@@ -16,6 +16,7 @@ package com.liferay.document.library.internal.exportimport.data.handler;
 
 import com.liferay.document.library.exportimport.data.handler.DLPluggableContentDataHandler;
 import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.model.DLFileEntryConstants;
 import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
 import com.liferay.document.library.kernel.model.DLFileEntryType;
 import com.liferay.document.library.kernel.model.DLFileVersion;
@@ -561,12 +562,20 @@ public class FileEntryStagedModelDataHandler
 						if (deleteFileEntry &&
 							ExportImportThreadLocal.isStagingInProcess()) {
 
-							_dlAppService.deleteFileVersion(
-								latestExistingFileVersion.getFileEntryId(),
-								latestExistingFileVersion.getVersion());
+							String latestExistingVersion =
+								latestExistingFileVersion.getVersion();
 
-							_overrideFileVersion(
-								importedFileEntry, fileVersion.getVersion());
+							if (!latestExistingVersion.equals(
+									importedFileEntry.getVersion())) {
+
+								_dlAppService.deleteFileVersion(
+									latestExistingFileVersion.getFileEntryId(),
+									latestExistingFileVersion.getVersion());
+
+								_overrideFileVersion(
+									importedFileEntry,
+									fileVersion.getVersion());
+							}
 						}
 					}
 					finally {
